requirements.txt
allure-pytest==2.13.5
pytest==8.2.1
selenium==4.21.0
webdriver-manager==4.0.1

